// src/interfaces/lesson.tsx
import {shedule} from '../utils/shedule';
// Generated by https://quicktype.io

export interface LessonsStateProps {
  lessons: Lesson[];
  schedlue: Schedlue[];
  day: string;
  saveLessons: (lessons: Lesson[]) => void;
  setSchedlue: (schedlue: Schedlue) => void;
  saveSchedule: (schedule: Schedlue[]) => void;
  deleteSchedlue: (schedule: Schedlue) => void;
  setDay: (day: string) => void;
  restoreSchedule: () => void;
  saveLesson: (lesson: Lesson) => void;
  deleteLessonStorage: (id: string) => void;
  editLessonStorage: (lesson: Lesson) => void;
}

export type LessonsActions =
  | {type: 'save lessons'; payload: {lessons: Lesson[]}}
  | {type: 'set schedlue'; payload: {schedlue: Schedlue}}
  | {type: 'save schedule'; payload: {schedule: Schedlue[]}}
  | {type: 'delete shedlue'; payload: {schedule: Schedlue}}
  | {type: 'set day'; payload: {day: string}}
  | {type: 'restore schedule'}
  | {type: 'save lesson'; payload: {lesson: Lesson}}
  | {type: 'delete lesson storage'; payload: {id: string}}
  | {type: 'edit lesson storage'; payload: {lesson: Lesson}};

export interface LessonsState {
  lessons: Lesson[];
  schedlue: Schedlue[];
  day: string;
}

export const lessonsInitialState: LessonsState = {
  lessons: [],
  schedlue: [],
  day: '',
};

export interface LessonResponse {
  lessons: Lesson[];
}

export interface Lesson {
  _id: string;
  lesson: string;
  nrc: number;
  teacher: string;
  schedlue: Schedlue[];
}

export interface Schedlue {
  _id?: number;
  day: string;
  hours: string;
}

export interface SaveLesson {
  lesson: string;
  nrc: string;
  schedlue: Schedlue[];
  teacher: string;
}